# Improved LoggedIO write performance