Note: the original idea in this PR does not work (well it does, but not the way I wanted it ;). ~~Alternatives are being developed.~~ Alternative developed, needs testing.

Also see https://github.com/borgbackup/borg/issues/45#issuecomment-213715080

Summary


Test set:

*  images + XMP metadata (for some folders)
* average file size ~3.5 MB
* 5388 files
* 19.49 GB

```
                        Throughput     %CPU      total CPU time
1.0.2 / master ........ 24 MB/s ...... 33 % .... 280 s
fastlio none .......... 52 MB/s ...... 62 % .... 237 s
fastlio none64 ........ 54 MB/s ...... 55 % .... 205 s
```

All are I/O-bound, but fastlio is less so. free(1) shows that master branch doesn't
let the kernel use it's buffers (due to fsync after each segment) efficiently, while
fastlio does. Profiling shows that for fastlio fsync() makes no significant wall-clock
appereance, while it is a large factor for master.

This patch doesn't (or shouldn't) trash the page cache, as it immediately tells the kernel after writing
that the data is not needed anymore. The fsync() comes much later, at a point where it seems
the kernel already wrote everything to disk on it's own accord. Hence these fsync()s return
immediately, instead of blocking.

While in my tests none and none64 did not make a tangible difference, none64 would be useful
where fast IO is available but the CPU isn't very fast (=NAS), since the average CPU loads
show it uses much less CPU time.

For machines running encrypted repositories with hardware AES support (AES-NI for example)
the throughput should scale similar. For machines with no hardware AES support the difference
will be less pronounced.