@enkore 2016-04-26
Detailed test log

```
$ borg init -e none64 liotest
$ sudo sync && echo 3 | sudo tee /proc/sys/vm/drop_caches
$ /bin/time borg create liotest::fastlio-none64 ~/largehome/Sync/Bilder -svp --no-files-cache
------------------------------------------------------------------------------
Archive name: fastlio-none64
Archive fingerprint: 24acf7bd996a25d8041356e1afd2297881e40a3b34026521d1cbaf245c6845ca
Time (start): Tue, 2016-04-26 14:45:48
Time (end):   Tue, 2016-04-26 14:51:53
Duration: 6 minutes 4.99 seconds
Number of files: 5388
------------------------------------------------------------------------------
                       Original size      Compressed size    Deduplicated size
This archive:               19.49 GB             19.49 GB             18.08 GB
All archives:               19.49 GB             19.49 GB             18.08 GB

                       Unique chunks         Total chunks
Chunk index:                   11162                12133
------------------------------------------------------------------------------
186.10user 18.45system 6:05.38elapsed 55%CPU (0avgtext+0avgdata 55384maxresident)k
38102378inputs+35359864outputs (0major+105642minor)pagefaults 0swaps

$ rm -rf liotest
$ borg init -e none liotest
$ sudo sync && echo 3 | sudo tee /proc/sys/vm/drop_caches
$ /bin/time borg create liotest::fastlio-none ~/largehome/Sync/Bilder -svp --no-files-cache
------------------------------------------------------------------------------
Archive name: fastlio-none
Archive fingerprint: 47ef43ac29025331dec62e515c75de92a55280bb6d94488d486c9eebedd1b8f7
Time (start): Tue, 2016-04-26 14:38:50
Time (end):   Tue, 2016-04-26 14:45:07
Duration: 6 minutes 17.43 seconds
Number of files: 5388
------------------------------------------------------------------------------
                       Original size      Compressed size    Deduplicated size
This archive:               19.49 GB             19.49 GB             18.08 GB
All archives:               19.49 GB             19.49 GB             18.08 GB

                       Unique chunks         Total chunks
Chunk index:                   11162                12133
------------------------------------------------------------------------------
219.13user 17.68system 6:18.18elapsed 62%CPU (0avgtext+0avgdata 56144maxresident)k
38144873inputs+35373736outputs (68major+84764minor)pagefaults 0swaps

$ ./borg-linux64 -V
borg-linux64 1.0.2
$ ./borg-linux64 init -e none liotest-master
$ sudo sync && echo 3 | sudo tee /proc/sys/vm/drop_caches
$ /bin/time ./borg-linux64 create liotest-master::master-none ~/largehome/Sync/Bilder -svp --no-files-cache
------------------------------------------------------------------------------
Archive name: master-none
Archive fingerprint: b6db9aaee8ff539120fff67bd678d5ca0f55574e2af131757879c4c568a4d08e
Time (start): Tue, 2016-04-26 12:49:03
Time (end):   Tue, 2016-04-26 13:02:55
Duration: 13 minutes 51.93 seconds
Number of files: 5388
------------------------------------------------------------------------------
                       Original size      Compressed size    Deduplicated size
This archive:               19.49 GB             19.49 GB             18.08 GB
All archives:               19.49 GB             19.49 GB             18.08 GB

                       Unique chunks         Total chunks
Chunk index:                   11165                12136
------------------------------------------------------------------------------
253.81user 26.69system 13:52.44elapsed 33%CPU (0avgtext+0avgdata 54912maxresident)k
38136530inputs+35385096outputs (3major+490703minor)pagefaults 0swaps
```