@ThomasWaldmann 2016-04-26
Can you keep the none64 changes in a different PR? Other than they are also tuning speed, they seem unrelated. See also the experimental repo with a lot of similar stuff.