@enkore 2016-04-26
I'll leave this here https://lwn.net/Articles/591723/ and return later.