@ThomasWaldmann 2016-04-27
non-commit writes: write immediately, queue fsync/fadvise (if queue is full, process first entry to make room).

commit writes: process queue entries (will be quick as kernel has already written most stuff anyway), write commit, fsync commit.