@enkore 2016-04-27
None of this works. A short queue (5-30 segments) doesn't make any difference, because the kernel just buffers it and doesn't start to write anything until the fsync()s come in, at which point it's back to stop-and-wait.

I then added (asynchronous) sync_file_range calls when enqueueing segments in the sync queue, which go through, but also doesn't motivate anything; all fdatasync calls still block for eternity, pretty much independent of queue length. Only when the queue length exceeds ~1 GB the performance gets up to where it was above -- with the exact same cache usage as before.

This will need to go in an entirely different direction to improve performance.