@ThomasWaldmann 2016-04-27
If we do not want to write and fsync-wait all the time, we just write and use the kernel's write-caching (and related policies / settings of the kernel).

If these settings allow a relatively big age of dirty write buffers (like >>1s), THIS is causing that a lot of dirty write buffers are pending. The more pending buffers, the longer we have to wait if we need to force them all on disk by fsyncing.

So I think the solution is not to allow such a big age of dirty write buffers on a system that writes big amounts of data. Modify the max age of dirty write buffers to a low value.