@enkore 2016-04-27
Yes, that should be the purpose of sync_file_range (it doesn't write any metadata, not even the meta necessary to find the data), to limit the amount of dirty pages, so that a later fdatasync completes much quicker. This does not have any effect here though. I am wondering whether this might be due to the small segment size (and the kernel thinking "what? syncing just 5 MB? nah, i don't think so"), I think I'll look into "jumbo-segments" as the next route, maybe that helps.