@ThomasWaldmann 2016-04-27
@enkore i wasn't talking about doing exotic calls in borg, but just setting that kernel setting to an appropriate value. That plus https://github.com/borgbackup/borg/pull/985#issuecomment-214938804 should solve it, right?