@enkore 2016-04-27
Hm, I didn't really consider those, since they are system-wide and affecting everything. On the other hand, they are more effective than any combination of syncing-queues, jumbo segments and/or sync_file_range I have tried so far. With that in mind, I am going to close this.

By the way, the defaults on my system equalled to "mere" 5 _gigabytes_ of dirty pages. Whoops.