@enkore 2016-04-27
Some points from @verygreen 

* Depending on file system fsync()s are not constrained to a single file, but more or less global = expensive
* Ideally just write write write as long as possible
* As few fsync()s as possible
* If the file system flushes it's transaction, then all data and metadata of the txn is consistent
* Segment files should not be too big, so that no extra indirection is needed to access contents. On ext4 this means 512 MiB.