@enkore 2016-05-08
Summary:

* Larger segments
* Smarter compaction (hints file is backwards incompatible, version=2)
* Commit mechanism as practised/recommended (double-sync-commit, first write data, then sync, then write commit tag, then sync again)
* Linux: sync_file_range for immediate write-out avoids fdatasync-stall (s_f_r *only* for performance, *not* for durability or anything like that)