@enkore 2016-05-09
Hole punching: Would now be possible with the changes from d6944df. Old versions would choke on it, though.