@ThomasWaldmann 2016-05-09
I would not update repo segment size setting automatically, rather document it.