@enkore 2016-05-09
> File systems

> We strongly recommend to not use Borg (or any other database-like software) on
> non-journaling file systems like FAT, since it is not possible to assume any
> consistency in case of power failures (or a sudden disconnect of an external drive).

> While Borg uses a data store that is resilient against these failures when used
> on journaling file systems, it is not possible to guarantee this with some hardware.